refactor(server): move handler assembly
