+++ internal/server/run.go
@@ -101,9 +101,3 @@ func shutdownAfterStartupError(srv *http.Server, timeout time.Duration, runErr e
 	}
 	return runErr
 }
-
-func handleHealthz(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	_, _ = w.Write([]byte(`{"status":"ok"}`))
-}
