+++ internal/server/security.go
@@ -1,7 +1,7 @@
 // Copyright 2026 the P&AI authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package main
+package server
 
 import (
 	"crypto/sha256"
